+++ examples/basic/server.prod.js
@@ -14,9 +14,10 @@ server.use(bodyParser.urlencoded({ extended: true }));
 
 server.use(express.static(path.join(__dirname, 'public')));
 
-server.use('/', (req, res) => ( // eslint-disable-line no-unused-vars
-  res.sendFile(path.join(__dirname, 'app', 'markup', 'bootstrap-prod.html'))
-));
+server.use('/', (
+  _req,
+  res, // eslint-disable-line no-unused-vars
+) => res.sendFile(path.join(__dirname, 'app', 'markup', 'bootstrap-prod.html')));
 
 server.listen(PORT, IP, err => {
   if (err) console.log(`=> OMG!!! 🙀 ${err}`);