+++ examples/basic/webpack.prod.config.js
@@ -4,50 +4,51 @@
 // https://github.com/shakacode/react-webpack-rails-tutorial/blob/master/client%2Fwebpack.client.base.config.js
 const webpack = require('webpack');
 
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const path = require('path');
-const autoprefixer = require('autoprefixer');
-const bootstrapEntryPoints = require('./webpack.bootstrap.config.js');
-
-// eslint-disable-next-line no-console
-console.log(`=> bootstrap-loader configuration: ${bootstrapEntryPoints.prod}`);
 
 module.exports = {
-
+  mode: 'production',
   // For production build we want to extract CSS to stand-alone file
   // Provide `extractStyles` param and `bootstrap-loader` will handle it
-  entry: [
-    'font-awesome-loader',
-    bootstrapEntryPoints.prod,
-    'tether',
-    './app/scripts/app',
-  ],
-
+  entry: ['tether', 'font-awesome-loader', 'bootstrap-loader', './app/scripts/app'],
   output: {
     path: path.join(__dirname, 'public', 'assets'),
     filename: 'app.js',
   },
-
-  resolve: { extensions: ['*', '.js'] },
-
   plugins: [
-    new ExtractTextPlugin({ filename: 'app.css', allChunks: true }),
+    new MiniCssExtractPlugin({ filename: 'app.css' }),
     new webpack.ProvidePlugin({
+      $: 'jquery',
+      jQuery: 'jquery',
+      'window.jQuery': 'jquery',
+      Tether: 'tether',
       'window.Tether': 'tether',
-    }),
-    new webpack.LoaderOptionsPlugin({
-      postcss: [autoprefixer],
+      Popper: ['popper.js', 'default'],
+      Alert: 'exports-loader?Alert!bootstrap/js/dist/alert',
+      Button: 'exports-loader?Button!bootstrap/js/dist/button',
+      Carousel: 'exports-loader?Carousel!bootstrap/js/dist/carousel',
+      Collapse: 'exports-loader?Collapse!bootstrap/js/dist/collapse',
+      Dropdown: 'exports-loader?Dropdown!bootstrap/js/dist/dropdown',
+      Modal: 'exports-loader?Modal!bootstrap/js/dist/modal',
+      Popover: 'exports-loader?Popover!bootstrap/js/dist/popover',
+      Scrollspy: 'exports-loader?Scrollspy!bootstrap/js/dist/scrollspy',
+      Tab: 'exports-loader?Tab!bootstrap/js/dist/tab',
+      Tooltip: 'exports-loader?Tooltip!bootstrap/js/dist/tooltip',
+      Util: 'exports-loader?Util!bootstrap/js/dist/util',
     }),
   ],
-
   module: {
     rules: [
-      { test: /\.css$/, use: ExtractTextPlugin.extract({
-        fallback: 'style-loader', use: 'css-loader!postcss-loader',
-      }) },
-      { test: /\.scss$/, use: ExtractTextPlugin.extract({
-        fallback: 'style-loader', use: 'css-loader!postcss-loader!sass-loader',
-      }) },
+      {
+        test: /\.css$/,
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'],
+      },
+
+      {
+        test: /\.scss$/,
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader', 'sass-loader'],
+      },
 
       {
         test: /\.woff2?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
@@ -59,15 +60,6 @@ module.exports = {
         test: /\.(ttf|eot|svg)(\?[\s\S]+)?$/,
         use: 'file-loader',
       },
-
-      // Use one of these to serve jQuery for Bootstrap scripts:
-
-      // Bootstrap 4
-      { test: /bootstrap\/dist\/js\/umd\//, use: 'imports-loader?jQuery=jquery' },
-
-      // Bootstrap 3
-      { test: /bootstrap-sass\/assets\/javascripts\//, use: 'imports-loader?jQuery=jquery' },
     ],
   },
-
 };