+++ examples/css-modules/app/components/CardColumns/CardColumns.jsx
@@ -10,21 +10,18 @@ export default class CardColumns extends React.Component {
         <div className="card-columns">
           <div className="card">
             <div className="card-block">
-              <h4 className="card-title">
-                Card title that wraps to a new line
-              </h4>
+              <h4 className="card-title">Card title that wraps to a new line</h4>
               <p className="card-text">
-                This is a longer card with supporting text
-                below as a natural lead-in to additional content.
-                This content is a little bit longer.
+                This is a longer card with supporting text below as a natural lead-in to additional
+                content. This content is a little bit longer.
               </p>
             </div>
           </div>
           <div className="card card-block">
             <blockquote className="card-blockquote">
               <p>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                Integer posuere erat a ante.
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere erat a
+                ante.
               </p>
               <footer>
                 <small className="text-muted">
@@ -37,8 +34,7 @@ export default class CardColumns extends React.Component {
             <div className="card-block">
               <h4 className="card-title">Card title</h4>
               <p className="card-text">
-                This card has supporting text below as a
-                natural lead-in to additional content.
+                This card has supporting text below as a natural lead-in to additional content.
               </p>
               <p className="card-text">
                 <small className="text-muted">Last updated 3 mins ago</small>
@@ -47,10 +43,7 @@ export default class CardColumns extends React.Component {
           </div>
           <div className="card card-block card-inverse card-primary text-xs-center">
             <blockquote className="card-blockquote">
-              <p>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                Integer posuere erat.
-              </p>
+              <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere erat.</p>
               <footer>
                 <small>
                   Someone famous in <cite title="Source Title">Source Title</cite>
@@ -61,20 +54,17 @@ export default class CardColumns extends React.Component {
           <div className="card card-block text-xs-center">
             <h4 className="card-title">Card title</h4>
             <p className="card-text">
-              This card has supporting text below as a
-              natural lead-in to additional content.
+              This card has supporting text below as a natural lead-in to additional content.
             </p>
             <p className="card-text">
-              <small className="text-muted">
-                Last updated 3 mins ago
-              </small>
+              <small className="text-muted">Last updated 3 mins ago</small>
             </p>
           </div>
           <div className="card">
             <blockquote className="card-blockquote">
               <p>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                Integer posuere erat a ante.
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere erat a
+                ante.
               </p>
               <footer>
                 <small className="text-muted">
@@ -86,8 +76,8 @@ export default class CardColumns extends React.Component {
           <div className="card card-block text-xs-right">
             <blockquote className="card-blockquote">
               <p>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit.
-                Integer posuere erat a ante.
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere erat a
+                ante.
               </p>
               <footer>
                 <small className="text-muted">
@@ -99,15 +89,12 @@ export default class CardColumns extends React.Component {
           <div className="card card-block">
             <h4 className="card-title">Card title</h4>
             <p className="card-text">
-              This is a wider card with supporting text below
-              as a natural lead-in to additional content.
-              This card has even longer content than the
-              first to show that equal height action.
+              This is a wider card with supporting text below as a natural lead-in to additional
+              content. This card has even longer content than the first to show that equal height
+              action.
             </p>
             <p className="card-text">
-              <small className="text-muted">
-                Last updated 3 mins ago
-              </small>
+              <small className="text-muted">Last updated 3 mins ago</small>
             </p>
           </div>
         </div>