+++ examples/multiple-entries/server.prod.js
@@ -14,19 +14,17 @@ server.use(bodyParser.urlencoded({ extended: true }));
 
 server.use(express.static(path.join(__dirname, 'public')));
 
-server.get('/',
-  (req, res) => res.redirect('/bs3'),
+server.get('/', (_req, res) => res.redirect('/bs3'));
+
+server.get('/bs3', (_req, res) =>
+  res.sendFile(path.join(__dirname, 'app', 'markup', 'bs3-prod.html')),
 );
 
-server.get('/bs3',
-  (req, res) => res.sendFile(path.join(__dirname, 'app', 'markup', 'bs3-prod.html')),
+server.get('/bs4', (_req, res) =>
+  res.sendFile(path.join(__dirname, 'app', 'markup', 'bs4-prod.html')),
 );
 
-server.get('/bs4',
-  (req, res) => res.sendFile(path.join(__dirname, 'app', 'markup', 'bs4-prod.html')),
-);
-
-server.listen(PORT, IP, (err) => {
+server.listen(PORT, IP, err => {
   if (err) console.log(`=> OMG!!! 🙀 ${err}`);
   console.log(`=> 🚀  Production server is running on port ${PORT}`);
 });