+++ src/utils/createBootstrapRequire.js
@@ -9,11 +9,8 @@ import path from 'path';
  * @returns {string}
  */
 export default function(module, bootstrapVersion, bootstrapPath) {
-  const scriptsPath = (
-    parseInt(bootstrapVersion, 10) === 3 ?
-    ['assets', 'javascripts', 'bootstrap'] :
-    ['js', 'dist']
-  );
+  const scriptsPath =
+    parseInt(bootstrapVersion, 10) === 3 ? ['assets', 'javascripts', 'bootstrap'] : ['js', 'dist'];
   const bootstrapModule = path.join(bootstrapPath, ...scriptsPath, module);
   return `require (${JSON.stringify(bootstrapModule)});`;
 }