+++ src/utils/logger.js
@@ -6,16 +6,13 @@ import chalk from 'chalk';
  * Logger helper
  */
 export default {
-
   /**
    * Logs output
    *
    * @param {...*} output
    */
   log(...output) {
-    const prettyOutput = (
-      [chalk.yellow('[bootstrap-loader]: ')].concat(output, '\n')
-    );
+    const prettyOutput = [chalk.yellow('[bootstrap-loader]: ')].concat(output, '\n');
     console.log(...prettyOutput);
   },
 
@@ -27,5 +24,4 @@ export default {
   debug(...output) {
     if (global.__DEBUG__) this.log(...output);
   },
-
 };